Navbar: refactors component location to navbar/ folder with images
